# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BioCypher Components Registry API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/adapters/search": {
      "get": {
        "tags": [
          "adapters"
        ],
        "summary": "Search public adapters",
        "description": "Return public adapter cards whose title matches the search term.",
        "operationId": "search_adapters_api_v1_adapters_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdapterLatestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adapters/{adapter_id}": {
      "get": {
        "tags": [
          "adapters"
        ],
        "summary": "Get adapter catalog detail",
        "description": "Return one public adapter. The response is metadata-light; fetch full Croissant metadata through the adapter metadata endpoint.",
        "operationId": "get_adapter_api_v1_adapters__adapter_id__get",
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            }
          },
          {
            "name": "bcr_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bcr Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdapterDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adapters/{adapter_id}/endorse": {
      "post": {
        "tags": [
          "adapters"
        ],
        "summary": "Endorse an adapter",
        "description": "Record that the signed-in user endorses this adapter.",
        "operationId": "endorse_adapter_api_v1_adapters__adapter_id__endorse_post",
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            }
          },
          {
            "name": "bcr_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bcr Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdapterEndorsementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adapters/{adapter_id}/metadata": {
      "get": {
        "tags": [
          "adapters"
        ],
        "summary": "Get adapter metadata",
        "description": "Return the full stored Croissant metadata document for one adapter.",
        "operationId": "get_adapter_metadata_api_v1_adapters__adapter_id__metadata_get",
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdapterMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdapterDataSourceResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "column_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column Count"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "AdapterDataSourceResponse",
        "description": "Metadata-light dataset summary embedded in adapter Croissant metadata."
      },
      "AdapterDetailResponse": {
        "properties": {
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id"
          },
          "adapter_name": {
            "type": "string",
            "title": "Adapter Name"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "repository_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Location"
          },
          "license_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Value"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "cff_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cff Url"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "maintainers": {
            "items": {
              "$ref": "#/components/schemas/AdapterMaintainerResponse"
            },
            "type": "array",
            "title": "Maintainers"
          },
          "data_sources": {
            "items": {
              "$ref": "#/components/schemas/AdapterDataSourceResponse"
            },
            "type": "array",
            "title": "Data Sources"
          },
          "endorsement_count": {
            "type": "integer",
            "title": "Endorsement Count",
            "default": 0
          },
          "endorsed_by_current_user": {
            "type": "boolean",
            "title": "Endorsed By Current User",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "adapter_id",
          "adapter_name"
        ],
        "title": "AdapterDetailResponse",
        "description": "Response model for one adapter."
      },
...
      "AdapterEndorsementResponse": {
        "properties": {
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id"
          },
          "endorsement_count": {
            "type": "integer",
            "title": "Endorsement Count"
          },
          "endorsed_by_current_user": {
            "type": "boolean",
            "title": "Endorsed By Current User",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "adapter_id",
          "endorsement_count"
        ],
        "title": "AdapterEndorsementResponse",
        "description": "Current endorsement state for one adapter."
      },
      "AdapterLatestItemResponse": {
        "properties": {
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id"
          },
          "adapter_name": {
            "type": "string",
            "title": "Adapter Name"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "repository_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Location"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "maintainers": {
            "items": {
              "$ref": "#/components/schemas/AdapterMaintainerResponse"
            },
            "type": "array",
            "title": "Maintainers"
          },
          "endorsement_count": {
            "type": "integer",
            "title": "Endorsement Count",
            "default": 0
          },
          "endorsed_by_current_user": {
            "type": "boolean",
            "title": "Endorsed By Current User",
            "default": false
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "adapter_id",
          "adapter_name",
          "updated_at"
        ],
        "title": "AdapterLatestItemResponse",
        "description": "Response model for one latest-adapters card."
      },
      "AdapterLatestListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdapterLatestItemResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AdapterLatestListResponse",
        "description": "Response model for newest adapter cards."
      },
      "AdapterMaintainerResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "AdapterMaintainerResponse",
        "description": "Repository maintainer identity for adapter catalog display."
      },
...
      "AdapterMetadataResponse": {
        "properties": {
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id"
          },
          "registry_entry_id": {
            "type": "string",
            "title": "Registry Entry Id"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "adapter_id",
          "registry_entry_id",
          "metadata"
        ],
        "title": "AdapterMetadataResponse",
        "description": "Response model for one adapter's full Croissant metadata."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}